[refactor] date util 함수 분리
